Isolate server code into an asyncio-specific module
